Expose `StorageTestSuite` struct for namespacing
