+++ identity-account-storage/src/storage/mod.rs
@@ -5,10 +5,12 @@ mod memstore;
 #[cfg(feature = "stronghold")]
 mod stronghold;
 #[cfg(feature = "storage-test-suite")]
-pub mod tests;
+mod test_suite;
 mod traits;
 
 pub use self::memstore::*;
 #[cfg(feature = "stronghold")]
 pub use self::stronghold::*;
 pub use self::traits::*;
+#[cfg(feature = "storage-test-suite")]
+pub use test_suite::StorageTestSuite;
